+++ tests/config.spec.ts
@@ -164,3 +164,50 @@ test(`
   await page.evaluate((microlc) => microlc.getApi().router.goToApplication('react'), microlcHandle)
   await expect(page.getByRole('link', { name: 'Go To About Page' })).toBeVisible()
 })
+
+test(`
+  [config injection]
+  parcel config as {"html": "<path>"}
+`, async ({ page }) => {
+  await goto(page, {
+    applications: {
+      react: {
+        entry: { html: 'https://micro-lc.io/applications/react-browser-router/' },
+        injectBase: true,
+        integrationMode: 'parcel',
+        route: './react/',
+      },
+    },
+    settings: {
+      defaultUrl: './react/',
+    },
+    version: 2,
+  })
+
+  const microlcHandle = await page.evaluateHandle<Microlc>('document.querySelector(\'micro-lc\')')
+
+  await page.evaluate((microlc) => microlc.getApi().router.goToApplication('react'), microlcHandle)
+  await expect(page.getByRole('link', { name: 'Go To About Page' })).toBeVisible()
+})
+
+test(`
+  [config injection]
+  error page customization should not override all defaults
+`, async ({ page }) => {
+  await goto(page, {
+    settings: {
+      defaultUrl: '/home',
+      errorPages: {
+        '4xx': {
+          400: {
+            integrationMode: 'iframe',
+            src: 'https://example.com',
+          },
+        },
+      },
+    },
+    version: 2,
+  })
+
+  await expect(page.locator('svg')).toBeVisible()
+})